diffmap.jl Documentation

Main.DiffusionMaps.DiffMap — Type

DiffMap{T <: Real} <: NonlinearDimensionalityReduction

The DiffMap type represents diffusion maps model constructed for T type data. It stores all the relevant information for the reduction.

StatsAPI.fit — Function

fit(::Type{DiffMap}, X::AbstractMatrix; 
    metric::PreMetric=DiffMetric(1.0), 
    maxoutdim::Int=2,
    ɛ::Real=1.0, 
    α::Real=0.5,
    t::Int=1)::DiffMap

Fit a DiffMap model to X using the specified parameters.

Arguments

::Type{DiffMap}: Type of the DiffMap model.

X::AbstractMatrix: Data matrix of observations. Each row of X is an observation.

Keyword arguments

metric::PreMetric: The metric used for computing the distance matrix. Default: DiffMetric(1.0)

maxoutdim::Int: The dimension of the reduced space. Default: 2

ɛ::Real: The scale parameter for the Gaussian kernel. Default: 1.0

α::Real: A normalization parameter. Default: 0.5

t::Int: The number of transitions. Default: 1

Returns

DiffMap: A DiffMap object representing the fitted model.

Examples

X = rand(100, 3) # toy data set, 100 observations with 3 features model = fit(DiffMap, X) # construct a DiffMap model R = projection(model) # Obtain the lower dimensional embedding

MultivariateStats.projection — Function

projection(M::DiffMap)

Returns the lower-dimensional embeddings of the input data. The format (rows = features or rows = observations) corresponds to the input format.

Arguments

M::DiffMap: The diffusion map model obtained from fitting.

StatsAPI.predict — Function

Embed new points into the reduced-dimensional space.

Arguments

model::DiffMap: The diffusion map model obtained from fitting.

new_points::AbstractMatrix: The new points to be embedded.

k::Int = 10: The number of nearest neighbors to consider.

This function embeds new points into the reduced-dimensional space of a DiffMap model. It therefore calculates the k nearest neigbours in the model's input data X and gives back the weightet average of their lower-dimensional embeddings.

Returns

An array containing the embedded representation of the new points.

Note

If the dimension of new_points is different from the dimension of the training data, a dimension mismatch error is raised.

The "newpoints" are not added to the model. The model stays unchanged. To alter the model and add the new points, consider calculating a new DiffMap model with newdata = hcat(model.X, newpoints) newmodel = fit(DiffMap, new_data)

Main.DiffusionMaps.metric — Function

metric(M)

Returns the metric used to calculate the kernel matrix of the model M. Defaults to the here defined DiffMetric.

Main.DiffusionMaps.kernel — Function

kernel(M)

Returns the kernel matrix of the model M. This is the normalized and symmetric distance matrix. The distance is calculated with the metric function. Call metric(M) to display the model's metric function.

Base.size — Function

size(M)

Returns a tuple with the dimensions of the reduced output (=d) and obervations (=the number of columns of the initial input).

MultivariateStats.indim — Function

indim(M)

Returns the input dimensions of the model.

MultivariateStats.outdim — Function

outdim(M)

Returns the output dimensions of the model.

LinearAlgebra.eigvals — Function

eigvals(M)

Return eigenvalues of the kernel matrix of the model M.

LinearAlgebra.eigvecs — Function

eigvecs(M)

Return eigenvectors of the kernel matrix of the model M.

Main.DiffusionMaps.normalize_laplacian — Function

normalize_laplacian(A::AbstractMatrix, α::Real)

Helper function within fit. Normalize the Laplacian matrix A using a power normalization factor α.

Arguments

A::AbstractMatrix: The Laplacian matrix to be normalized.

α::Real: The power normalization factor.

Returns

res::AbstractMatrix: The normalized Laplacian matrix.

This function computes the normalized Laplacian matrix by applying a power normalization factor α to the input matrix A.

It first constructs a diagonal matrix D by summing the rows of A and placing the resulting sums on the diagonal.

Then, it applies the power normalization operation D^(-α) * A * D^(-α) to obtain the normalized Laplacian matrix.

Main.DiffusionMaps.normalize_direction! — Function

normalize_direction!(A::AbstractMatrix)

Helper function within fit. Adjust the direction of columns in the matrix A.

Arguments

A::AbstractMatrix: The matrix whose column directions need to be normalized.

This function adjusts the direction of each column in the matrix A. It iterates over the columns of A and checks if the first element of each column is negative. If so, it flips the sign of all elements in that column, effectively normalizing its direction. This makes sure, that the fitting of the same data has the same lower dimensional embedding and doesn't differ because of the eigenvector's sign.

Note: This function modifies the input matrix A in place.

Main.DiffusionMaps.DiffMetric — Type

Structure for the here used distance function