Diffusion Maps

Diffusion Maps is a similarity preserving dimensionality reduction.

DiffMap Struct

The DiffMap type represents a Diffusion Maps model constructed for T type data. It stores all the relevant information for the reduction.

struct DiffMap{T <: Real} <: NonlinearDimensionalityReduction
    X::AbstractMatrix{T}
    d::Number
    t::Int
    ɛ::Real
    metric::PreMetric
    λ::AbstractVector{T}
    V::AbstractMatrix{T}
    P::AbstractMatrix{T}
end

Main.DiffusionMaps.DiffMap — Type

DiffMap{T <: Real} <: NonlinearDimensionalityReduction

The DiffMap type represents diffusion maps model constructed for T type data. It stores all the relevant information for the reduction.

DiffMetric Struct

Structure for the distance function used in the Diffusion Maps. It is the implementation of the Gaussian Kernel.

struct DiffMetric <: PreMetric 
    ɛ::Real
end

Main.DiffusionMaps.DiffMetric — Type

Structure for the here used distance function

Properties

size(M): Returns a tuple with the dimensions of the reduced output (=d) and observations (=the number of columns of the initial input).

Base.size — Function

size(M)

Returns a tuple with the dimensions of the reduced output (=d) and obervations (=the number of columns of the initial input).

indim(M): Returns the input dimensions of the model.

MultivariateStats.indim — Function

indim(M)

Returns the input dimensions of the model.

outdim(M): Returns the output dimensions of the model.

MultivariateStats.outdim — Function

outdim(M)

Returns the output dimensions of the model.

eigvals(M): Return eigenvalues of the kernel matrix of the model M.

LinearAlgebra.eigvals — Function

eigvals(M)

Return eigenvalues of the kernel matrix of the model M.

eigvecs(M): Return eigenvectors of the kernel matrix of the model M.

LinearAlgebra.eigvecs — Function

eigvecs(M)

Return eigenvectors of the kernel matrix of the model M.

metric(M): Returns the metric used to calculate the kernel matrix of the model M. Defaults to the here defined DiffMetric.

Main.DiffusionMaps.metric — Function

metric(M)

Returns the metric used to calculate the kernel matrix of the model M. Defaults to the here defined DiffMetric.

kernel(M): Returns the kernel matrix of the model M. This is the normalized and symmetric distance matrix. The distance is calculated with the metric function.

Main.DiffusionMaps.kernel — Function

kernel(M)

Returns the kernel matrix of the model M. This is the normalized and symmetric distance matrix. The distance is calculated with the metric function. Call metric(M) to display the model's metric function.

Functions

Exported Functions

fit(::Type{DiffMap}, X::AbstractMatrix; ...)::DiffMap: Fit a DiffMap model to X using specified parameters.

StatsAPI.fit — Function

fit(::Type{DiffMap}, X::AbstractMatrix; metric::PreMetric=DiffMetric(1.0), maxoutdim::Int=2, ɛ::Real=1.0, α::Real=0.5, t::Int=1, sparse_enabled::Bool=false, tol::Real=1e-6)::DiffMap

Fit a DiffMap model to X using the specified parameters.

Arguments

::Type{DiffMap}: Type of the DiffMap model.

X::AbstractMatrix: Data matrix of observations. Each row of X is an observation.

Keyword arguments

metric::PreMetric: The metric used for computing the distance matrix. Default: DiffMetric(1.0)

maxoutdim::Int: The dimension of the reduced space. Default: 2

ɛ::Real: The scale parameter for the Gaussian kernel. Default: 1.0

α::Real: A normalization parameter. Default: 0.5

t::Int: The number of transitions. Default: 1

sparse_enabled::Bool: Enable sparse representation of the kernel matrix. Default: false

tol::Real: Tolerance for sparse representation. Values below tol in the similarity matrix will be set to zero. Default: 1e-6

Returns

DiffMap: A DiffMap object representing the fitted model.

Examples

X = rand(100, 3) # toy data set, 100 observations with 3 features model = fit(DiffMap, X) # construct a DiffMap model R = projection(model) # Obtain the lower dimensional embedding

projection(M::DiffMap): Returns the lower-dimensional embeddings of the input data.

MultivariateStats.projection — Function

projection(M::DiffMap)

Returns the lower-dimensional embeddings of the input data. The format (rows = features or rows = observations) corresponds to the input format.

Arguments

M::DiffMap: The diffusion map model obtained from fitting.

predict(model::DiffMap, data::AbstractMatrix; k::Int = 1): Embed new points into the reduced-dimensional space.

StatsAPI.predict — Function

predict(model::DiffMap, data::AbstractMatrix; k::Int = 1)

Embed new points into the reduced-dimensional space.

Arguments

model::DiffMap: The diffusion map model obtained from fitting.

new_points::AbstractMatrix: The new points to be embedded.

k::Int = 10: The number of nearest neighbors to consider.

This function embeds new points into the reduced-dimensional space of a DiffMap model. It therefore calculates the k nearest neigbours in the model's input data X and gives back the weightet average of their lower-dimensional embeddings.

Returns

An array containing the embedded representation of the new points.

Note

If the dimension of new_points is different from the dimension of the training data, a dimension mismatch error is raised.

The "newpoints" are not added to the model. The model stays unchanged. To alter the model and add the new points, consider calculating a new DiffMap model with newdata = hcat(model.X, newpoints) newmodel = fit(DiffMap, new_data)

reconstruct(model::DiffMap, data::AbstractMatrix; k::Int = 1): Calculate an approximation of the higher-dimensional input, given the lower-dimensional output.

MultivariateStats.reconstruct — Function

reconstruct(model::DiffMap, data::AbstractMatrix; k::Int = 1)

Calculate an approximation of the higher dimensional imput, given the lower dimensional output.

Arguments

model::DiffMap: The diffusion map model obtained from fitting.

data::AbstractMatrix: The data points to be reconstructed in the original input space.

k::Int = 10: The number of nearest neighbors to consider for reconstruction.

This functioncalculates an approximated higher dimensional input based on a DiffMap model. It calculates the k nearest neighbors in the model's input data X and returns the weighted average of their higher-dimensional embeddings.

Returns

An array containing the reconstructed representation of the new points in the original input space.

Note

If the dimension of data is different from the dimension of the projection data, a dimension mismatch error is raised.

The "data" points are not added to the model. The model remains unchanged. To alter the model and add the new points, consider calculating a new DiffMap model with new_data = hcat(model.X, data) new_model = fit(DiffMap, new_data)

Helper Functions

normalize_laplacian(S::AbstractMatrix, α::Real)::AbstractMatrix: Helper function within fit to normalize the Laplacian matrix.

Main.DiffusionMaps.normalize_laplacian — Function

normalize_laplacian(A::AbstractMatrix, α::Real)

Helper function within fit. Normalize the Laplacian matrix A using a power normalization factor α.

Arguments

A::AbstractMatrix: The Laplacian matrix to be normalized.

α::Real: The power normalization factor.

Returns

res::AbstractMatrix: The normalized Laplacian matrix.

This function computes the normalized Laplacian matrix by applying a power normalization factor α to the input matrix A.

It first constructs a diagonal matrix D by summing the rows of A and placing the resulting sums on the diagonal.

Then, it applies the power normalization operation D^(-α) * A * D^(-α) to obtain the Laplacian matrix.

Lastly it computes the now-wise normalization D^(-α) * L.

normalize_direction!(A::AbstractMatrix): Helper function within fit to adjust the direction of columns in the matrix A.

Main.DiffusionMaps.normalize_direction! — Function

normalize_direction!(A::AbstractMatrix)

Helper function within fit. Adjust the direction of columns in the matrix A.

Arguments

A::AbstractMatrix: The matrix whose column directions need to be normalized.

This function adjusts the direction of each column in the matrix A. It iterates over the columns of A and checks if the first element of each column is negative. If so, it flips the sign of all elements in that column, effectively normalizing its direction. This makes sure, that the fitting of the same data has the same lower dimensional embedding and doesn't differ because of the eigenvector's sign.

Note: This function modifies the input matrix A in place.